(CONT-785) Rubocop Auto Fixes 6-10
- Lint/DeprecatedOpenSSLConstant
- RSpec/EmptyLineAfterExample
- RSpec/ReceiveNever
- Style/FetchEnvVar
- Style/IfUnlessModifier

+++ lib/puppet/provider/java_ks/keytool.rb
@@ -164,9 +164,7 @@ Puppet::Type.type(:java_ks).provide(:keytool) do
     rescue => e
       if e.message.match?(%r{password was incorrect}i)
         # we have the wrong password for the keystore. so delete it if :password_fail_reset
-        if @resource[:password_fail_reset] == :true
-          File.delete(@resource[:target])
-        end
+        File.delete(@resource[:target]) if @resource[:password_fail_reset] == :true
       end
       false
     end
